Pulling presentation logic out of core libs
